requirements.txt
python-telegram-bot[webhooks]
starlette
flask
supabase
python-dotenv
fpdf2
qrcode
Pillow
requests
aiohttp
pytz
certifi
yt-dlp
openai
shazamio
audioop-lts
pydantic
pydantic-settings
redis
ujson